DEV-15: Cleaning the code and final changes
