+++ src/pages/CamerasPage/Footer/CamerasFooter.module.scss
@@ -10,30 +10,3 @@
     padding: 0.5rem;
     background-color: colors.getColor("primary", 30);
 }
-
-// .cameraFooter :nth-child(1) {
-//     // flex-grow: 1;
-//     // background-color: purple;
-// }
-
-// .info {
-//     // flex-grow: 1;
-//     width: 5rem;
-
-//     height: 100%;
-//     display: flex;
-//     flex-direction: column;
-//     align-items: center;
-//     justify-content: center;
-//     border-style: solid;
-//     border-color: black;
-//     border-width: 1px;
-//     border-radius: 10px;
-//     padding: 10px;
-// }
-
-// .tube {
-//     flex-grow: 1;
-//     // background-color: purple;
-//     // height: 100%;
-// }
