+++ src/pages/CamerasPage/Footer/CamerasFooter.tsx
@@ -1,8 +1,6 @@
-//import React from "react";
-//import "./Footer.css";
 import styles from "./CamerasFooter.module.scss";
-import { Information } from "./Information";
-import { Tube } from "./Tube";
+import { Information } from "./Information/Information";
+import { Tube } from "./Tube/Tube";
 
 export const CamerasFooter = () => {
     return (
