+++ src/pages/CamerasPage/Footer/Information/Information.module.scss
@@ -11,17 +11,6 @@
 
     color: colors.getColor("primary", 80);
     font-size: fonts.getFontSize("x-small");
-
-    // height: 100%;
-    // display: flex;
-    // flex-direction: column;
-    // align-items: center;
-    // justify-content: center;
-    // border-style: solid;
-    // border-color: black;
-    // border-width: 1px;
-    // border-radius: 10px;
-    // padding: 10px;
 }
 
 .measurementLine {
