+++ src/pages/CamerasPage/Footer/Tube/Kenos/Kenos.module.scss
@@ -1,11 +1,9 @@
 @use "src/styles/colors.scss";
 .kenos {
-    //flex-grow: 1;
     width: 15%;
     min-width: 4rem;
     max-width: 8rem;
     height: 2.5rem;
-    // display: flex;
     position: absolute;
     left: 30px;
