+++ src/pages/CamerasPage/Footer/Tube/Kenos/Kenos.tsx
