+++ src/pages/CamerasPage/SecondaryCameras/SecondaryCameras.module.scss
@@ -1,0 +1,13 @@
+@use "src/styles/colors.scss";
+.secondaryCameras {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    width: 30%;
+    gap: 1rem;
+    min-width: 8rem;
+    max-height: 19rem;
+    max-width: 16rem;
+}
